requirements/base.txt
Django 
PyMuPDF 
djangorestframework
